+++ InternetBank.Infrastructure/Migrations/20241025070204_AddAccount.cs
@@ -24,7 +24,7 @@ namespace InternetBank.Infrastructure.Migrations
                     ExpireDate = table.Column<DateTime>(type: "timestamp with time zone", nullable: false),
                     accountType = table.Column<int>(type: "integer", nullable: false),
                     AccountStaticPassword = table.Column<string>(type: "text", nullable: false),
-                    Amount = table.Column<int>(type: "integer", nullable: false)
+                    Amount = table.Column<long>(type: "bigint", nullable: false)
                 },
                 constraints: table =>
                 {
